started re-organising mopper and mop_setup, added setup and init files
