+++ setup.cfg
@@ -1,0 +1,38 @@
+[metadata]
+name = mopper 
+url = https://github.com/ACCESS-Hive/ACCESS-MOPPeR
+author = Paola Petrelli, Sam Green
+author_email = paola.petrelli@utas.edu.au, sam.green@unsw.edu.au
+summary = 'ACCESS Model Output Post-Processor, maps raw model output to CMIP-style defined variables and produce post-processed output using CMOR3' 
+description_file = README.md
+licence = 'Apache-2.0'
+keywords = 'ACCESS model'
+classifier =
+    Development Status :: 3 - Alpha
+    Environment :: Console
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: Apache Software License
+    Operating System :: POSIX :: Linux
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9 
+    Programming Language :: Python :: 3.10 
+
+
+[files]
+packages = 
+    mopper 
+
+[pbr]
+autodoc_tree_index_modules = True
+autodoc_tree_excludes =
+    setup.py
+    test
+    docs/conf.py
+
+[build_sphinx]
+source_dir = docs
+build_dir = docs/_build
+
+[tool:pytest]
+addopts = --doctest-modules --doctest-glob='*.rst' --ignore setup.py --ignore docs/conf.py
